% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-group_by.R
\name{group_data}
\alias{group_data}
\alias{group_data.PlySummarizedExperiment}
\title{get grouping data}
\usage{
\method{group_data}{PlySummarizedExperiment}(.data)
}
\arguments{
\item{.data}{An object Inheriting from \code{PlySummarizedExperiment}, the wrapper
class for \code{SummarizedExperiment} objects}
}
\value{
list of groupings for an SummarizedExperiment
}
\description{
retrieve grouping information from a \code{SummarizedExperiment} object. This
is stored within the \code{metadata()} of the object.
}
\examples{
group_by(se_simple, rows(direction), cols(condition)) |> group_data()
}
