% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_utils.R
\name{group_vars}
\alias{group_vars}
\alias{group_vars.PlySummarizedExperiment}
\title{get PlySummarizedExperiment grouping Variables}
\usage{
\method{group_vars}{PlySummarizedExperiment}(x)
}
\arguments{
\item{x}{PlySummarizedExperiment}
}
\value{
NULL or list containing names of grouping columns
}
\description{
like in \code{\link[dplyr:group_vars]{dplyr::group_vars()}} will get character strings for
groupings with the exection of the return value
being a list for each grouped context
}
\examples{

out <- group_by(se_simple, rows(direction))
group_vars(out)

}
