% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-S4-unchop.R
\name{list_unchop}
\alias{list_unchop}
\title{unchop a list of objects}
\usage{
list_unchop(x, ptype = NULL, ..., indices = NULL)
}
\arguments{
\item{x}{a list}

\item{ptype}{the expected prototype of the output}

\item{...}{unused arguments}

\item{indices}{optional list of integer vectors whose size is equal to that
of x. This maps the the final index of each element in the output.}
}
\value{
an object of type \code{ptype} or the common ptype of elements of x.
}
\description{
A generic version of \code{\link[vctrs:list_unchop]{vctrs::list_unchop}} meant to
support S4 Vectors.
}
