% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyxp-class.R
\name{se}
\alias{se}
\alias{se,PlySummarizedExperiment-method}
\alias{se<-}
\alias{se<-,PlySummarizedExperiment-method}
\alias{se,SummarizedExperiment-method}
\alias{se<-,SummarizedExperiment-method}
\alias{PlySummarizedExperiment-methods}
\alias{assays,PlySummarizedExperiment-method}
\alias{assays<-,PlySummarizedExperiment,list-method}
\alias{assays<-,PlySummarizedExperiment,SimpleList-method}
\alias{assay,PlySummarizedExperiment,missing-method}
\alias{assay,PlySummarizedExperiment,numeric-method}
\alias{assay,PlySummarizedExperiment,character-method}
\alias{assay<-,PlySummarizedExperiment,missing-method}
\alias{assay<-,PlySummarizedExperiment,numeric-method}
\alias{assay<-,PlySummarizedExperiment,character-method}
\alias{rowData,PlySummarizedExperiment-method}
\alias{rowData<-,PlySummarizedExperiment-method}
\alias{colData,PlySummarizedExperiment-method}
\alias{colData<-,PlySummarizedExperiment,DataFrame-method}
\alias{colData<-,PlySummarizedExperiment,NULL-method}
\alias{metadata,PlySummarizedExperiment-method}
\alias{metadata<-,PlySummarizedExperiment-method}
\alias{rownames,PlySummarizedExperiment-method}
\alias{colnames,PlySummarizedExperiment-method}
\alias{nrow,PlySummarizedExperiment-method}
\alias{ncol,PlySummarizedExperiment-method}
\alias{dimnames,PlySummarizedExperiment-method}
\alias{dimnames<-,PlySummarizedExperiment,list-method}
\alias{dimnames<-,PlySummarizedExperiment,NULL-method}
\title{PlySummarizedExperiment Methods}
\usage{
se(x)

\S4method{se}{PlySummarizedExperiment}(x)

se(x) <- value

\S4method{se}{PlySummarizedExperiment}(x) <- value

\S4method{se}{SummarizedExperiment}(x)

\S4method{se}{SummarizedExperiment}(x) <- value

\S4method{assays}{PlySummarizedExperiment}(x, withDimnames = TRUE, ...)

\S4method{assays}{PlySummarizedExperiment,list}(x, withDimnames = TRUE, ...) <- value

\S4method{assays}{PlySummarizedExperiment,SimpleList}(x, withDimnames = TRUE, ...) <- value

\S4method{assay}{PlySummarizedExperiment,missing}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{PlySummarizedExperiment,numeric}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{PlySummarizedExperiment,character}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{PlySummarizedExperiment,missing}(x, i, withDimnames = TRUE, ...) <- value

\S4method{assay}{PlySummarizedExperiment,numeric}(x, i, withDimnames = TRUE, ...) <- value

\S4method{assay}{PlySummarizedExperiment,character}(x, i, withDimnames = TRUE, ...) <- value

\S4method{rowData}{PlySummarizedExperiment}(x, use.names = TRUE, ...)

\S4method{rowData}{PlySummarizedExperiment}(x, ...) <- value

\S4method{colData}{PlySummarizedExperiment}(x, ...)

\S4method{colData}{PlySummarizedExperiment,DataFrame}(x, ...) <- value

\S4method{colData}{PlySummarizedExperiment,NULL}(x, ...) <- value

\S4method{metadata}{PlySummarizedExperiment}(x, ...)

\S4method{metadata}{PlySummarizedExperiment}(x, ...) <- value

\S4method{rownames}{PlySummarizedExperiment}(x)

\S4method{colnames}{PlySummarizedExperiment}(x)

\S4method{nrow}{PlySummarizedExperiment}(x)

\S4method{ncol}{PlySummarizedExperiment}(x)

\S4method{dimnames}{PlySummarizedExperiment}(x)

\S4method{dimnames}{PlySummarizedExperiment,list}(x) <- value

\S4method{dimnames}{PlySummarizedExperiment,NULL}(x) <- value
}
\arguments{
\item{x}{PlySummarizedExperiment object}

\item{value}{replacement value}

\item{withDimnames}{logical}

\item{...}{additional arguments}

\item{i}{character or numeric index}

\item{use.names}{logical}
}
\value{
Replacement functions return a PlySummarizedExperiment object. Other functions will return the same object as the method from SummarizedExperiment.
}
\description{
Methods from SummarizedExperiment package re-implemented for PlySummarizedExperiment.
}
\section{Functions}{
\itemize{
\item \code{se(PlySummarizedExperiment)}: get the se slot of the PlySummarizedExperiment object

\item \code{se(x) <- value}: set the se slot of the PlySummarizedExperiment object

\item \code{se(PlySummarizedExperiment) <- value}: set the se slot of the PlySummarizedExperiment object

\item \code{se(SummarizedExperiment)}: get the SummarizedExperiment object

\item \code{se(SummarizedExperiment) <- value}: get the SummarizedExperiment object

\item \code{assays(PlySummarizedExperiment)}: get the assays o the PlySummarizedExperiment object

\item \code{assays(x = PlySummarizedExperiment) <- value}: set the assays of the PlySummarizedExperiment object

\item \code{assays(x = PlySummarizedExperiment) <- value}: set the assays of the PlySummarizedExperiment object

\item \code{assay(x = PlySummarizedExperiment, i = missing)}: get the first assay of the PlySummarizedExperiment object

\item \code{assay(x = PlySummarizedExperiment, i = numeric)}: get assay from a PlySummarizedExperiment object

\item \code{assay(x = PlySummarizedExperiment, i = character)}: get assay from a PlySummarizedExperiment object

\item \code{assay(x = PlySummarizedExperiment, i = missing) <- value}: set assay in a PlySummarizedExperiment object

\item \code{assay(x = PlySummarizedExperiment, i = numeric) <- value}: set assay in a PlySummarizedExperiment object

\item \code{assay(x = PlySummarizedExperiment, i = character) <- value}: set assay in a PlySummarizedExperiment object

\item \code{rowData(PlySummarizedExperiment)}: get rowData in a PlySummarizedExperiment object

\item \code{rowData(PlySummarizedExperiment) <- value}: set rowData in a PlySummarizedExperiment object

\item \code{colData(PlySummarizedExperiment)}: get colData in a PlySummarizedExperiment object

\item \code{colData(x = PlySummarizedExperiment) <- value}: set colData in a PlySummarizedExperiment object

}}
\examples{
assays(se_simple)
rowData(se_simple)
colData(se_simple)

}
