% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyxp-package.R
\docType{package}
\name{plyxp-package}
\alias{plyxp-package}
\title{plyxp: Data masks for SummarizedExperiment enabling dplyr-like manipulation}
\value{
API for using S4 classes with rlang data masks
}
\description{
The package provides `rlang` data masks for the SummarizedExperiment class. The enables the evaluation of unquoted expression in different contexts of the SummarizedExperiment object with optional access to other contexts. The goal for `plyxp` is for evaluation to feel like a data.frame object without ever needing to unwind to a rectangular data.frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jtlandis/plyxp}
  \item \url{https://jtlandis.github.io/plyxp}
  \item Report bugs at \url{https://www.github.com/jtlandis/plyxp/issues}
}

}
\author{
\strong{Maintainer}: Justin Landis \email{jtlandis314@gmail.com} (\href{https://orcid.org/0000-0001-5501-4934}{ORCID})

Authors:
\itemize{
  \item Justin Landis \email{jtlandis314@gmail.com} (\href{https://orcid.org/0000-0001-5501-4934}{ORCID})
  \item Michael Love \email{michaelisaiahlove@gmail.com} (\href{https://orcid.org/0000-0001-8401-0545}{ORCID})
}

}
\keyword{internal}
