% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-slice_min.R
\name{slice_min}
\alias{slice_min}
\alias{slice_min.PlySummarizedExperiment}
\title{Slice minimum rows or columns of a PlySummarizedExperiment}
\usage{
\method{slice_min}{PlySummarizedExperiment}(
  .data,
  ...,
  n = 1,
  prop = NULL,
  with_ties = TRUE,
  na_rm = FALSE,
  .preserve = FALSE
)
}
\arguments{
\item{.data}{a PlySummarizedExperiment object}

\item{...}{expressions wrapped in \code{rows()} or \code{cols()} giving the ordering
variable. Only the \code{rows()} and \code{cols()} contexts are available here.}

\item{n}{Number of rows/columns to select per group. Cannot be combined
with \code{prop}.}

\item{prop}{Fraction of rows/columns to select per group. Cannot be
combined with \code{n}.}

\item{with_ties}{Should ties be kept together? Default \code{TRUE}.}

\item{na_rm}{Should missing values be removed? Default \code{FALSE}.}

\item{.preserve}{When \code{FALSE} (default), grouping structure is recomputed
based on the resulting data.}
}
\description{
\code{slice_min()} selects rows or columns with the smallest values of a variable.
}
\examples{
slice_min(se_simple, rows(length), n = 2)

gse <- group_by(se_simple, rows(direction))
slice_min(gse, rows(length), n = 1)
}
