% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-S4-size.R
\name{vctrs-vec_size}
\alias{vctrs-vec_size}
\alias{vec_size}
\title{replicate a vector}
\usage{
vec_size(x)
}
\value{
size of the S3 or S4 vector
}
\description{
A re-export of \code{\link[vctrs:vec_size]{vctrs::vec_size}} and
as an S7 generic function to allow \code{S4Vectors}.
}
\examples{
vec_size(1:2, times = 5)
vec_size(S4Vectors::Rle(1:2), times = 5)
}
