% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-S4-unchop.R
\name{vec_ptype2}
\alias{vec_ptype2}
\title{Create a common prototype of two vectors}
\usage{
vec_ptype2(x, y, ...)
}
\arguments{
\item{x}{first object}

\item{y}{second object}

\item{...}{unused arguments}
}
\description{
given two objects, \code{vec_ptype2()} finds a common prototype
}
