% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-S4-unchop.R
\name{vec_ptype_common}
\alias{vec_ptype_common}
\title{find the common ptype}
\usage{
vec_ptype_common(..., .ptype = NULL)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> a collection of objects}

\item{.ptype}{the expected prototype}
}
\value{
an object of size 0.
}
\description{
find the common ptype
}
