% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-utils.R
\docType{data}
\name{dot-pronouns}
\alias{dot-pronouns}
\alias{.assays}
\alias{.assays_asis}
\alias{.rows}
\alias{.rows_asis}
\alias{.cols}
\alias{.cols_asis}
\title{contextual plyxp pronouns}
\format{
An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.
}
\usage{
.assays

.assays_asis

.rows

.rows_asis

.cols

.cols_asis
}
\value{
access to specific values behind the rlang pronoun
}
\description{
\code{plyxp} utilizes its own version of \code{rlang::.data} pronouns. These may be
used to gain access to other evaluation contexts for a managed set of
data-masks.

Similar to \code{rlang::.data}, \code{plyxp::.assays} and other exported pronouns
are exported to pass R CMD Checks. When using a \code{plyxp} within your package,
import the associated pronoun from \code{plyxp} but only use the fully unqualified
name, \code{.assays}, \code{.assays_asis}, etc.
}
\examples{
mutate(
  se_simple,
  # access via pronoun
  rows(sum = rowSums(.assays_asis$counts)),
  cols(sum = vapply(.assays$counts, sum, numeric(1)))
)

}
\keyword{datasets}
