% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyxp-class.R
\name{new_plyxp}
\alias{new_plyxp}
\alias{PlySummarizedExperiment}
\title{SummarizedExperiment Shell Object}
\usage{
new_plyxp(se)

PlySummarizedExperiment(se = class_missing)
}
\arguments{
\item{se}{SummarizedExperiment object}
}
\value{
PlySummarizedExperiment object
}
\description{
A container object for the SummarizedExperiment class.

This is an S7 class only exported for possible extension.
}
\examples{
new_plyxp(se = se_simple@se)
}
