% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyxp-class.R
\docType{class}
\name{new_plyxp}
\alias{new_plyxp}
\alias{PlySummarizedExperiment-class}
\alias{PlySummarizedExperiment}
\title{SummarizedExperiment Shell Object}
\usage{
new_plyxp(se)
}
\arguments{
\item{se}{SummarizedExperiment object}
}
\value{
PlySummarizedExperiment object
}
\description{
A container object for the SummarizedExperiment class.

This is an S4 class only exported for possible extension.
}
\examples{
se <- SummarizedExperiment(
  assays = list(counts = matrix(1:6, nrow = 3)),
  colData = DataFrame(condition = c("A", "B"))
)
new_plyxp(se = se)
# or
PlySummarizedExperiment(se = se)
}
