% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyxp-class.R
\name{plyxp}
\alias{plyxp}
\title{Modify SummarizedExperiment Object}
\usage{
plyxp(.data, .f, ...)
}
\arguments{
\item{.data}{a PlySummarizedExperiment object}

\item{.f}{a function that returns a SummarizedExperiment object}

\item{...}{additional arguments passed to \code{.f}}
}
\value{
a PlySummarizedExperiment object
}
\description{
Modify the underlying SummarizedExperiment object with a function.
}
\examples{
plyxp(se_simple, function(x) x)
}
