% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-dplyr.R
\name{reexports}
\alias{reexports}
\title{reexports from other packages}
\value{
exported functions available from \code{plyxp}
}
\description{
reexports from other packages
}
\examples{
arrange(se_simple, rows(direction)) |>
  mutate(logp_counts = log1p(counts)) |>
  filter(cols(condition == "drug"))

assays(se_simple)
rowData(se_simple)
colData(se_simple)

}
\seealso{
\code{\link[=arrange]{arrange()}} \code{\link[=mutate]{mutate()}} \code{\link[=filter]{filter()}}
\code{\link[=summarize]{summarize()}} \code{\link[=select]{select()}} \code{\link[=pull]{pull()}} \code{\link[=group_by]{group_by()}}
\code{\link[=group_data]{group_data()}} \code{\link[=group_vars]{group_vars()}} \code{\link[=ungroup]{ungroup()}}

\link{PlySummarizedExperiment-methods}
}
