% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-SummarizedExperiment.R,
%   R/reexports-dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{SummarizedExperiment}
\alias{assay}
\alias{assay<-}
\alias{assays}
\alias{assays<-}
\alias{rowData}
\alias{rowData<-}
\alias{colData}
\alias{colData<-}
\alias{metadata}
\alias{metadata<-}
\title{Objects exported from other packages}
\value{
exported functions available from \code{plyxp}
}
\examples{
arrange(se_simple, rows(direction)) |>
  mutate(logp_counts = log1p(counts)) |>
  filter(cols(condition == "drug"))

assays(se_simple)
rowData(se_simple)
colData(se_simple)

}
\seealso{
\code{\link[=arrange]{arrange()}} \code{\link[=mutate]{mutate()}} \code{\link[=filter]{filter()}}
\code{\link[=summarize]{summarize()}} \code{\link[=select]{select()}} \code{\link[=pull]{pull()}} \code{\link[=group_by]{group_by()}}
\code{\link[=group_data]{group_data()}} \code{\link[=group_vars]{group_vars()}} \code{\link[=ungroup]{ungroup()}}

\link{PlySummarizedExperiment-methods}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{S4Vectors}{\code{\link[S4Vectors:Annotated-class]{metadata}}, \code{\link[S4Vectors:Annotated-class]{metadata<-}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}}
}}

