% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-slice.R
\name{slice}
\alias{slice}
\alias{slice.PlySummarizedExperiment}
\title{Slice a PlySummarizedExperiment}
\usage{
\method{slice}{PlySummarizedExperiment}(.data, ..., .preserve = FALSE)
}
\description{
\code{slice()} selects rows and/or columns by position.
Slice Documentation TO DO
}
