% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-S4-group.R
\name{vec_group_id}
\alias{vec_group_id}
\alias{vec_group_loc}
\title{Create Group Indices}
\usage{
vec_group_id(x, ...)

vec_group_loc(x, ...)
}
\arguments{
\item{x}{A vector}
}
\value{
either a tibble or DataFrame object
}
\description{
A alternative to \code{\link[vctrs:vec_group]{vctrs::vec_group_id}} and
\code{\link[vctrs:vec_group]{vctrs::vec_group_loc}} as
an S4 generic function to allow \code{S4Vectors}.
}
\examples{

data <- S4Vectors::DataFrame(
  letter = sample(letters, 500, TRUE),
  LETTER = sample(letters, 500, TRUE)
)

vec_group_id(data)
vec_group_loc(data)

}
