% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-S4-.R
\docType{data}
\name{vectors}
\alias{vectors}
\alias{class_vctrs}
\alias{class_s4_vctrs}
\alias{class_DF}
\title{S7 classes for vctrs and S4 Vectors}
\format{
An object of class \code{S7_union} of length 1.

An object of class \code{classRepresentation} of length 1.

An object of class \code{classRepresentation} of length 1.
}
\usage{
class_vctrs

class_s4_vctrs

class_DF
}
\value{
S7 class union or base class
}
\description{
A set of S7 classes and Class unions that help establish S7 method dispatch.
These classes were made to re-export several \code{vctrs} functions such that
internals for \code{plyxp} were consistent with room for optimization.
}
\examples{
# used for defining methods on S7 generics

S7::method(vec_slice, class_s4_vctrs)

}
\seealso{
\code{\link[=vec_rep]{vec_rep()}},\code{\link[=vec_recycle]{vec_recycle()}},\code{\link[=vec_slice]{vec_slice()}}
}
\keyword{datasets}
